Moving Health to other Spring Libraries. Renamed Client modifier, Removed old test.
