+++ sdk/appconfiguration/spring-cloud-azure-appconfiguration-config/src/main/java/com/azure/spring/cloud/config/implementation/config/AppConfigurationAutoConfiguration.java
@@ -29,7 +29,7 @@ public class AppConfigurationAutoConfiguration {
      */
     @Configuration
     @ConditionalOnClass(RefreshEndpoint.class)
-    static class AppConfigurationWatchAutoConfiguration {
+    public static class AppConfigurationWatchAutoConfiguration {
 
         @Bean
         @ConditionalOnMissingBean
