+++ sdk/spring/spring-cloud-azure-actuator-autoconfigure/src/main/java/com/azure/spring/cloud/actuator/autoconfigure/appconfiguration/AppConfigurationConfigHealthConfiguration.java
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package com.azure.spring.cloud.config.implementation.config;
+package com.azure.spring.cloud.actuator.autoconfigure.appconfiguration;
 
 import org.springframework.boot.actuate.autoconfigure.health.ConditionalOnEnabledHealthIndicator;
 import org.springframework.boot.actuate.health.HealthIndicator;
@@ -10,9 +10,9 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
+import com.azure.spring.cloud.actuator.appconfiguration.AppConfigurationConfigHealthIndicator;
 import com.azure.spring.cloud.config.AppConfigurationRefresh;
 import com.azure.spring.cloud.config.implementation.config.AppConfigurationAutoConfiguration.AppConfigurationWatchAutoConfiguration;
-import com.azure.spring.cloud.config.implementation.health.AppConfigurationHealthIndicator;
 
 /**
  * Health Indicator for Azure App Configuration store connections.
@@ -21,12 +21,12 @@ import com.azure.spring.cloud.config.implementation.health.AppConfigurationHealt
 @ConditionalOnClass({ HealthIndicator.class })
 @ConditionalOnEnabledHealthIndicator("azure-app-configuration")
 @AutoConfigureAfter(AppConfigurationWatchAutoConfiguration.class)
-public class AppConfigurationHealthAutoConfiguration {
+public class AppConfigurationConfigHealthConfiguration {
 
     @Bean
     @ConditionalOnBean(AppConfigurationRefresh.class)
-    AppConfigurationHealthIndicator appConfigurationHealthIndicator(AppConfigurationRefresh refresh) {
-        return new AppConfigurationHealthIndicator(refresh);
+    AppConfigurationConfigHealthIndicator appConfigurationHealthIndicator(AppConfigurationRefresh refresh) {
+        return new AppConfigurationConfigHealthIndicator(refresh);
     }
 
 }
